Debugger disaster
